Reference PNG version of Caltech logo via absolute URL
When releasing software on PyPI, using SVG for the icon does not
render at all.  Further, the URL has to be an absolute URL instead of
a relative URL into the `.graphics/` subdirectory.  Prior to now, I
would always fix this in every individual project I worked on, but
that's stupid.  Let's just fix the template.