Remove section on installing modules to avoid duplication
The info about installing required modules is usually included in the
README of a project, and the information is usually written in terms
of using a `requirements.txt` file.  I think it's better not to have
another set of instructions in this file, especially since it's not
actually about installing Python per se.